+ Add main() function test MyPoint class

+++ PointCordination/PointCordination/MyPoint.h
@@ -15,6 +15,6 @@ public:
 	int GetX();
 	void SetY(int);
 	int GetY();
-	float Distance(MyPoint p);
+	double Distance(MyPoint p);
 };
 #endif
